backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-jose[cryptography]
pydantic-settings